chore: introduce ruff
