+++ pyproject.toml
@@ -26,10 +26,43 @@ build-backend = "hatchling.build"
 
 [tool.rye]
 managed = true
-dev-dependencies = []
+dev-dependencies = [
+    "ruff>=0.5.1",
+]
 
 [tool.hatch.metadata]
 allow-direct-references = true
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/quizscatterer"]
+packages = ["quizscatterer"]
+
+# Lint/Formatter非依存のルール
+[tool.ruff]
+# 除外したいファイル
+# excludeを設定するとデフォルトで設定されている.venv配下などの設定が上書きされるため要注意
+extend-exclude = ["src/quizscatterer/__init__.py"]
+# 行文字数
+line-length = 88
+# インデント幅
+indent-width = 4
+
+# Linterのルール
+[tool.ruff.lint]
+# Lintルール
+# https://docs.astral.sh/ruff/rules/　に記載されているルールのうち、Codeが前方一致するルールが適用されます
+select = ["E", "F", "I", "PLR", "PD", "NPY", "PL"]
+# 除外するLintルール
+# selectで選択したルールから除外したいものを指定できます
+ignore = ["PLR2004"]
+# auto fixの対象にするルール
+fixable = ["ALL"]
+# auto fixから除外するルール
+# editor.formatOnSaveを使う際に自動でFixされると困るものを指定できます
+unfixable = ["F401"]
+
+# Formatterのルール
+[tool.ruff.format]
+# 文字列をダブルクォートに統一
+quote-style = "double"
+# インデントをタブ文字(\t)からスペースに変換
+indent-style = "space"
