+++ requirements-dev.lock
@@ -25,6 +25,7 @@ python-dateutil==2.9.0.post0
 pytz==2024.1
     # via pandas
     # via quizscatterer
+ruff==0.5.1
 scipy==1.14.0
     # via gensim
     # via quizscatterer
