+++ setup.py
@@ -2,9 +2,9 @@ from setuptools import setup
 
 setup(
     name="quizScatterer",
-    version='1.0',
-    description='word2vecとクラスタリングでクイズの出題順を最適化したい',
-    author='Miyax',
-    url='https://github.com/miyax0227/quizScatterer',
-    install_requires=open('requirements.txt').read().splitlines(),
+    version="1.0",
+    description="word2vecとクラスタリングでクイズの出題順を最適化したい",
+    author="Miyax",
+    url="https://github.com/miyax0227/quizScatterer",
+    install_requires=open("requirements.txt").read().splitlines(),
 )
