--- src/quizscatterer/__init__.py
@@ -1,2 +1,0 @@
-def hello() -> str:
-    return "Hello from quizscatterer!"
